[#14] refactor: move mail and template to utils files
